Add files via upload
Improved data gathering code. Now the device waits for the phone to send Start message before starting to gather and send data. If the connection if unexpectedly severed the device stops processing and sending. When the connection is reestablished the device continues the data processing and sending. The phone needs to send a Stop message for the device to fully stop processing.
With this Bluetooth connection can be maintained without the device processing and sending data